+++ setup.py
@@ -10,7 +10,5 @@ setup(
     packages=find_packages(include=["metagraph_cuda", "metagraph_cuda.*"]),
     include_package_data=True,
     install_requires=["metagraph", "cudf", "cugraph"],
-    entry_points={
-        "metagraph.plugins": "plugins=metagraph_cuda.registry:find_plugins"
-    },
+    entry_points={"metagraph.plugins": "plugins=metagraph_cuda.plugins:find_plugins"},
 )
